why is it such a nightmare to load fonts in SFML?
+ Text on buttons
+ Increase/Decrease playback speed buttons
+ Buttons positioned and centered correctly... also a bit of a nightmare somehow.

+++ CellularAutomataCustom/Main.cpp
@@ -9,7 +9,7 @@ void main()
 	sf::Clock clock;
 
 	//delta time between frames -- to be passed into all update functions from here.
-	float deltaTime = clock.restart().asMilliseconds();
+	float deltaTime = (float)clock.restart().asMilliseconds();
 	sf::Clock deltaClock;
 
 	//enable vsync
