+++ CellularAutomataCustom/SceneManager.cpp
@@ -13,12 +13,12 @@ SceneManager::~SceneManager()
 
 void SceneManager::PreUpdate(double a_DeltaTime)
 {
-	m_TileScene2d->PreUpdate(a_DeltaTime);
+	m_TileScene2d->PreUpdate((float)a_DeltaTime);
 }
 
 void SceneManager::Update(double a_DeltaTime)
 {
-	m_TileScene2d->Update(a_DeltaTime);
+	m_TileScene2d->Update((float)a_DeltaTime);
 }
 
 void SceneManager::Draw()
